implemented basic error handling
implemented read/write/query.

+++ Sources/Globals.swift
@@ -16,19 +16,19 @@ public struct DefaultSchemas {
         Action(createTable: "ShardLocation"),
         Action(addColumn: "shardId", type: .String, table: "ShardLocation"),
         Action(addColumn: "node", type: .String, table: "ShardLocation"),
-        Action(addColumn: "available", type: .Numeric, table: "ShardLocation"),
+        Action(addColumn: "available", type: .Int, table: "ShardLocation"),
         
         Action(createTable: "Node"),
         Action(addColumn: "locationId", type: .String, table: "Node"),
         Action(addColumn: "nodegroup", type: .String, table: "Node"),
-        Action(addColumn: "available", type: .Numeric, table: "Node"),
-        Action(addColumn: "last_seen", type: .Numeric, table: "Node"),
+        Action(addColumn: "available", type: .Int, table: "Node"),
+        Action(addColumn: "last_seen", type: .Int, table: "Node"),
         Action(addColumn: "internalAddress", type: .String, table: "Node"),
-        Action(addColumn: "internalPort", type: .Numeric, table: "Node"),
+        Action(addColumn: "internalPort", type: .Int, table: "Node"),
         Action(addColumn: "externalAddress", type: .String, table: "Node"),
-        Action(addColumn: "externalPort", type: .Numeric, table: "Node"),
-        Action(addColumn: "spaceUsed", type: .Numeric, table: "Node"),
-        Action(addColumn: "spaceFree", type: .Numeric, table: "Node"),
+        Action(addColumn: "externalPort", type: .Int, table: "Node"),
+        Action(addColumn: "spaceUsed", type: .Int, table: "Node"),
+        Action(addColumn: "spaceFree", type: .Int, table: "Node"),
         
     ]
